Bug #1463810 Qogirl6: Solving WCN module compilation problems
[root cause  ] Qogirl6 need to Solve WCN module compilation problems
[changes     ] Solving WCN module compilation problems
[side effects] no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:no
[is there dependence]:no
[confirm dependent commit]:yes
[board]:ums9230_1h10_go2g
[test case]:pass
[reviewers   ] baojie.cai
[change_type ] bringup --> sprd_chip_support
[tag_product ] common
Test: make sprdwl_ng.ko ok!

+++ wlan/marlin3_pcie/if_sc2355.c
@@ -1681,7 +1681,7 @@ int sprdwl_tx_free_pcie_data(struct sprdwl_priv *priv, unsigned char *data)
 	struct sprdwl_intf *dev = (struct sprdwl_intf *)priv->hw_priv;
 	struct sprdwl_tx_msg *tx_msg = (struct sprdwl_tx_msg *)dev->sprdwl_tx;
 	void *data_addr_ptr;
-	unsigned long pcie_addr = 0;
+	unsigned long long pcie_addr = 0;
 	unsigned long phy_addr;
 	unsigned short	data_num;
 	struct txc_addr_buff *txc_addr;
